refactored package name to neutronium, builds

+++ cmd/neutronium-compress/nw.go
@@ -134,9 +134,9 @@ func nwAlign(rseq, oseq []byte, mem *memory) [2][]byte {
 
 func alignmentDistance(alignment [2][]byte) float64 {
 	hammingDist := float64(0)
-	aLen := len(alignment[0])
+	aLen := float64(len(alignment[0]))
 
-	for i := aLen - 1; i >= 0; i-- {
+	for i := int(aLen - 1); i >= 0; i-- {
 		if alignment[0][i] != alignment[1][i] {
 			hammingDist += 1
 		}
